requirements.txt
pyrogram==2.0.106
TgCrypto
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.0
aiohttp==3.9.3
